Add missing SupportedOSPlatform attributes

+++ sources/Interop/Windows/Windows/shared/ws2ipdef/SIO.cs
@@ -8,20 +8,20 @@ namespace TerraFX.Interop.Windows;
 public static partial class SIO
 {
     [NativeTypeName("#define SIO_GET_INTERFACE_LIST _IOR('t', 127, ULONG)")]
-    public const int SIO_GET_INTERFACE_LIST = (0x40000000 | (((int)(4) & 0x7f) << 16) | (('t') << 8) | (127));
+    public const int SIO_GET_INTERFACE_LIST = (0x40000000 | (((int)(4) & 0x7f) << 16) | (((sbyte)('t')) << 8) | (127));
 
     [NativeTypeName("#define SIO_GET_INTERFACE_LIST_EX _IOR('t', 126, ULONG)")]
-    public const int SIO_GET_INTERFACE_LIST_EX = (0x40000000 | (((int)(4) & 0x7f) << 16) | (('t') << 8) | (126));
+    public const int SIO_GET_INTERFACE_LIST_EX = (0x40000000 | (((int)(4) & 0x7f) << 16) | (((sbyte)('t')) << 8) | (126));
 
     [NativeTypeName("#define SIO_SET_MULTICAST_FILTER _IOW('t', 125, ULONG)")]
-    public const uint SIO_SET_MULTICAST_FILTER = unchecked(0x80000000 | (((int)(4) & 0x7f) << 16) | (('t') << 8) | (125));
+    public const uint SIO_SET_MULTICAST_FILTER = unchecked(0x80000000 | (((int)(4) & 0x7f) << 16) | (((sbyte)('t')) << 8) | (125));
 
     [NativeTypeName("#define SIO_GET_MULTICAST_FILTER _IOW('t', 124 | IOC_IN, ULONG)")]
-    public const uint SIO_GET_MULTICAST_FILTER = unchecked(0x80000000 | (((int)(4) & 0x7f) << 16) | (('t') << 8) | (124 | 0x80000000));
+    public const uint SIO_GET_MULTICAST_FILTER = unchecked(0x80000000 | (((int)(4) & 0x7f) << 16) | (((sbyte)('t')) << 8) | (124 | 0x80000000));
 
     [NativeTypeName("#define SIO_IDEAL_SEND_BACKLOG_QUERY _IOR('t', 123, ULONG)")]
-    public const int SIO_IDEAL_SEND_BACKLOG_QUERY = (0x40000000 | (((int)(4) & 0x7f) << 16) | (('t') << 8) | (123));
+    public const int SIO_IDEAL_SEND_BACKLOG_QUERY = (0x40000000 | (((int)(4) & 0x7f) << 16) | (((sbyte)('t')) << 8) | (123));
 
     [NativeTypeName("#define SIO_IDEAL_SEND_BACKLOG_CHANGE _IO('t', 122)")]
-    public const int SIO_IDEAL_SEND_BACKLOG_CHANGE = (0x20000000 | (('t') << 8) | (122));
+    public const int SIO_IDEAL_SEND_BACKLOG_CHANGE = (0x20000000 | (((sbyte)('t')) << 8) | (122));
 }
